[app-mapper] Respond with 503s while the app is spinning up
When proxing a request for a new app, the app-mapper accesses 'GET /api' in the
app to confirm whether it's ready.

While the app is not ready, app-mapper will inmediately respond with a 503.

The readiness of the app is stored in the database to ensure we only ping the
app until it reaches a ready state.