Fix Prometheus scrape targets
Prometheus will get instances by doing a DNS A query on e.g.
users.default.svc.cluster.local. When it was backed by WeaveDNS, this
would return all instances; when it is backed by Kubernetes+SkyDNS, this
returns the singular service IP.

This change uses a new-ish feature in Prometheus, which queries the API
server to deduce the list of targets.

Should address #274, need to run it it prod to be sure.