Use the git state in an image tag, and push only that
This tags each image built by the Makefiles with the git state (as
calculated in the script `image-tag`), including the commit ID, the
branch, and whether there are uncommitted changes in the working
directory.

The script `tag-and-push` is replaced by `push-images`, which attempts
to push images for the components given as arguments (same as
`tag-and-push`), with the tag of each image constructed from the git
state as above.

The idea is that it is difficult to inadvertently push images that
aren't what you think they are, e.g., by building then changing
branches.

The exception is "WIP" builds, i.e., when the working directory has
uncommitted changes. These are all tagged with a trailing "WIP" no
matter what the changes, so it's possible to push images that do not
correspond to the state of the working directory. Try not to do that.