Add option to push-images to push only changed images
When building the Docker images in CI, every image is built from
scratch and pushed. But in most commits, only one or two images will
actually have different content.

This would be OK, except that pushing fresh images takes a long
time. Using `push-images --if-changed-since` makes the script push only
the images which have changes in the last commit.

In the absence and presumed difficulty of a more general solution,
"changed" means "a difference in a file under the directory
corresponding to the image". To make this more explicit, and things
more uniform, components are now referred to by a directory name (the
image name is assumed to be the base name).

Lastly, the CircleCI config has been adapted to look for the last
successful build on the branch, and use that as the value for
`--if-changed-since`.