Fixes scrollbar visibility issues. (#596)
The default "hide scrollbars when not scrolling" allowed us to be a bit
more relaxed about our overflow code. When a user opts to "always show
scrollbars" (a browser or OS setting), we're showing too many
scrollbars. Re-org our overflows a bit so even in this case we don't
show too many scrollbars. Scrollbars.