Disable idle connection pooling in authfe
Should increase latency, but improve load-balancing. I believe that the
connection pooling is interacting badly with iptables-based virtual IPs,
causing most of the traffic to be routed to one host. In particular,
golang uses a stack for it's idle connection pool, so that one host will
get almost all the connections.