Give nginx explicit hash bucket size
Nginx creates a hash table for server names based on the cache line
size it detects. Usually this ends up being 64; but, in some virtual
machines (one using KVM, in my case) it incorrectly defaults to
32. This is not big enough for the server names used in frontend's
configuration, and nginx refuses to start.

To work around this, set it to the required value (64) explicitly.