authfe: /demo is 404ing, fix it by removing redirect entirely
The problem is that /demo/ is transformed to /demo and passed to PathPrefix,
so we can't actually idependently match /demo and /demo/ under our current scheme!
So we get rid of the redirect entirely and just have /demo and /demo/ both go to the same place.
We do this by fixing the regex to also match /demo (previously it wasn't, so we were passing the
/demo path through, getting a 404).