Add flux v6 argument and routes
So that we can roll the new Flux API out without removing the prior
versions, this commit adds specific routes for v6 of that API, and an
argument for the hostname to which to forward requests.

To avoid having to change config exactly in sync with this code being
deployed, the new routes just use the hostname from the original
argument, and ignore the value of the new argument. Once config has
been updated to supply the new argument, they can be differentiated in
the code (and thereafter, the config).