Add GetDelinquentOrganizations() to users gRPC (#1352)
* Add GetDelinquentOrganizations() to users gRPC

Also adds database columns `zuora_account_number` and
`zuora_account_created_at` to the `organizations` table to keep track of
whether an organization has added a payment method or not. This allows
us to determine whether an organization is delinquent or not.

* DB organization selects/updates should exclude deleted

The unique constraint for an organization's external_id is only over the
set of rows that have `deleted_at is NULL`. We therefore should only filter
for these rows when doing comparison.

* Make linter/tests happy

* Fix swallowed db error when generating external id

* Fix zuora account stuff not being saved

Also, return NotFound for deleted instance.

* Introduce ExternalIDUsed() to cover deleted ids

Also,
- We now return 404 for a deleted organization (previously 403)
- Removed most of the `lower()` pgsql method calls on `external_id`
  since we already lower when inserting, no need to do so when reading
- Organization DB methods generally act like deleted orgs do not exist
  with the exception of `ExternalIDUsed()` which also includes those
- Replaces unique constraint over non-deleted orgs with a constraint
  over all organization, reuse of externalID is not permitted

* Add missing index name suffix `_idx`