Add utilities to randomly enable feature flags (+tests).
This introduces the `Enabler` abstraction, which can be used for arbitrary logic to decide to enable/disable feature flags.

It is admittedly under-used and over-engineered at this point, the only method being `IsEnabled` and the only implementation being `RandomEnabler` (which is a simple condition on a pseudo-random number generator), but the goal is to encourage using such abstractions to encapsulate logic related to feature flags.
Feel free to make this abstraction evolve to, for example, also include calls to the DB to update feature flags, etc.