Don't set debug log mode in every test
Removed code that called it on every test - if you need debug info you
will have to re-insert that line on the one test you are interested in.

Originally this was done because logging.Setup() could not be called
more than once; that restriction was eased upstream but it still seems
reasonable to leave logging off unless you actually need it.