circleci: fix deployment step
deployment steps for different branches need to be included in different sub-sections.
From the docs:
`Subsection names must be unique. Each subsection can list multiple branches, but at least one of these fields must be named branch. In instances of multiple branches, the first one that matches the branch being built is the one that is run.`
and
`Subsection names must be unique. Each subsection must list at least one branch or tag. In the case of multiple subsections, the first one that matches the branch being built is the only one that is run.`