Add `withPlatform` boolean arg to flux status API
If `withPlatform` is `false`, the flux API service does not attempt to
populate the elements of the status structure that involve making RPC
calls to the platform, giving a method to quickly ascertain connected
status and last connection time without the risk of length RPC timeouts.

If not supplied at all, `true` is assumed for backwards compatibility.