Fix http.Render() returning 400 instead of 500
`MalformedInputError` is an alias to the `error` interface and
will therefore match with every other error type in a `err.(type)`
switch statement.

This PR introduces conrete types to detect them properly.

Additionally, pointer receivers are used in order for object
comparison to return inequality unless it is the very same object.