Enable GCP subscription approval
It was previously disabled because we have a limited set of GCP Billing
Accounts which can each only have a single subscription. Once that
subscription is approved, we can no longer go through the signup flow.

One such account is hardcoded not to approve it, all others will execute
an approval call.