CircleCi deployment step: compare against master
push-images compares the codebase against the previous commit in the current branch in circleci.
If circleci does not know the previous commit it returns null. There's no point in checking for differences againt null.
Comparing always against master allows to reproduce the build's behaviour in master from a non-master branch.