Give admins a button to remove user from org
Occasionally we may need to intervene to remove a (bogus) account from
an instance. This commit adds a means of doing this through the admin
interface.

It's already possible to remove a user from an instance by becoming a
user in that instance; however, one of the reasons we may need to
intervene is if this doesn't work for whatever reason.

The admin version of this function uses the ID of the user rather than
the email address, since the latter might be the cause of the problem.