Use flux git repo status to inform setup message
We now receive a readiness status from the flux daemon while it's
starting up, so we can say with more precision what may be lacking.

In addition, we can use that to avoid calling `SyncStatus` for new
daemons, which will report a spurious error if the daemon has failed
to sync successfully yet.

For older daemons, we have to fall back to using `SyncStatus`; but we
can at least convey what has happened in more congruent terms.