Deprecate and remove code for flux API v1-v5
We have been running the old service in parallel, so these routes are
not used other than in a local environment.

Once we change authfe (config first, then code) in dev and prod to
route all requests to flux-api, we will have to deal with v1-v5
here. This commit prepares for that by explicitly deprecating those
versions, responding w9th `410 Gone` and an error suggesting an
upgrade.

In the case of the daemon endpoints, fluxd v0.2.0 will attempt to
reconnect more or less continuously; for that reason, I effectively
limit the rate at which they can do so by not responding for five
seconds. fluxd v0.3 and up will log the message and exit if the daemon
endpoint responds `410 Gone`.