authfe: Support Bearer token authentication for push endoints
Bearer tokens are quite widely used for API authentications. The first word of
the Authorization header is then "Bearer" and most APIs don't expect a base64
encoded token in this case.

Prometheus can use a bearer token for remote_writes. We'd like to use
bearer_token_file as it allows us to store the Weave Cloud token in a k8s
Secret and expose it as a file in the prometheus pod.

https://prometheus.io/docs/prometheus/1.8/configuration/configuration/#<remote_write>