Remove differentiated flux v6 routes
Flux API up to v5 is now deprecated, and all flux routes can point to
the new service (which will handle all the requests correctly).

This commit removes the differentiated routes, and the extra host
argument `-flux-v6`, in favour of all routes pointing to the new
service, as supplied in the argument `-flux`.

The argument `-flux-v6` will still be accepted (in case it lingers in
config), but only `-flux` will be used.