Fix cycle start if cycle day is inexistant day of month
We should never use a cycle day that's not within the first week anyway.
But this commit fixes the case for a billing cycle day of 31 while we
are computing the start of the next cycle when looking at the last day
of a month with 30 days.

NextCycleStart(2017-11-30, 31) did previously return 2017-11-30 and it
now returns 2017-12-31.