Add modifiable "trial remaining" admin column
This adds an input field to the admin organization list that displays
the remaining days in a organization's trial. It can be modified and
saved; optionally with an email that goes out to all the members of the
organization to notify about the trial extension. It is only extendable
but cannot be shrunk at the moment.