Allow UpdateOrganization to set notify ts to NULL
When we extend the trial period, we want to have another email sent out
once the trial expires, even if that has already been sent out at the
end of the last trial end.

This change allows us to reset the notification date to NULL and
therefore make the email trigger again.