Close incoming Github webhook connections after payload validation
An `http.Handler` will keep the connection backing its `http.ResponseWriter`
open at least until the point where the handler returns. The handler in
github-receiver would previously not return until its request to flux-api
completed.

In the situation where the specified instance's fluxd is not connected to
flux-api, the result is a 10 second wait followed by a timeout and an error
response which Github interprets as a delivery failure.

Validating the payload is sufficient information to let Github know delivery was
successful, so the remainder of processing must not block the handler.

Fixes #1890