Remove `ql` DB driver from flux-api
This driver used to be used for running tests that rely on a database without
having to run postgres. Since we now run all such tests against postgres, this
is safe to remove.

Also included in this commit is the addition of a build tag to some tests which
fit the above description, but which were previously run in the unit test suite.