Remove unnecessary struct-wrapping of eventsURL
There exists no codepath which assigns a value that is not known at compile-time
to any field in `instance.Config` before proceeding to pass that config as a
parameter to `notifications.Event`. As such, the existing code just pointlessly
wraps the URL to which we send events in a struct which can be deleted.

This commit also removes the `Username` field from `notifications.slackMsg`
because it is never assigned, along with the `GetConfig`, `SetConfig`, and
`PatchConfig` methods on `*server.Server` as they are unused.