Use the Result field in flux events, not the Spec
The Spec field gives what was intended to happen, but this does not
always correspond to what happened. For example, automated releases
can sometimes over-estimate the changes required for a set of new
images, if files are updated in the meantime.

Instead, use the Results field, which records what was successfully
updated.

This commit also changes the format of Commit notifications so that
the revision is mentioned once (there is always exactly one).