Convert flux-api to standard flag package
Everything else in weaveworks/service & weaveworks/common (and then onto
cortex etc) uses the standard package; replacing all those uses with
spf13/pflag will generate a lot of churn. The downside is we lose the
`-l` short flag for listen, but we don't use that in any of our
manifests anyway.