Do not allow joining any team by ext id
There are two ways to create an instance with team. One is assigning a
team by name. This checks whether an accessible team has that name and
if not, creates a new one.

The second way is assigning a team by its teamExternalID. This check
verifies correctly whether an user is already part of it. But if they
aren't, we look for any team by that ID and add the user to that team.

This allows anyone to add themselves to any team by just knowing its
external id and therefore gaining access to _all_ the other instances in
that team.