Revendor weaveworks/flux to get Antecedent field
Soon to be released flux daemons mark services/controllers/workloads
as having an "antecedent" if they were created via the Helm operator
telling Helm to release a chart. So we can develop UI against this,
update the flux API definitions used by flux-api so as to include this
field, by revendoring weaveworks/flux.