Restore webhook body after consuming to check sig
Verifying the signature of the webhook payload necessarily reads the
request body, so before we pass it to the next handler, we need to
give the request a replacement body.

NB: the procedure `github.ValidatePayload(...)` is supposed to return
the payload (suitably decoded); however, we're not going to use that
here, since we want to faithfully proxy the request rather than
interpret it.