Use logger object from common/server
If the main program and the library both register a flag `-log.level`
then the program will abort.  In general we standardise on the one
from package `weaveworks/common/server`, except where the program did
not register that package's flags and used its own.

Add logging initialization to a couple of programs that had none, and
thus non-standard formatting.