Handle deleted GCP accounts when processing pubsub
If the account referenced by the external account ID is deleted on
Google's side we receive an event referencing the subscriptions.
When we then try to fetch those subscriptions they are no longer
available and we return 500, making the PubSub message to be sent over
and over again.

For accounts that are deleted, we already received events that marked
the subscrpitions as complete.

So in the case of not being able to find a subscription, and that
subscription is locally marked as complete, we can just ACK the message.