Ignore empty GCP pubsub messages
We are currently receiving empty messages which we NACK and as such will
be resent over and over. I presume this might be because Google runs
code that supports Subscriptions API (legacy) and Procurement API (new)
at the same time.

Functionality for subscribing and changing entitlement plans is there so
we can just ignore them for now and wait for Google to respond to the
inquiry.