Revert gogo/protobuf upgrade
When we updated to use go 1.10.x, we also updated the version of
gogo/protobuf in an attempt to work around a dependency issue. This in
turn caused the generated users.pb.go to change, which is probably
undesirable if we can avoid it. Solve the problem in a different way.