Ensure WriteError is given a JSONable error
There are two funcs provided by flux's HTTP package (here imported as
`transport`): `ErrorResponse` and `WriteError`. The latter is slightly
lower-level, and is used here for things that aren't application
errors, e.g., for problems parsing the request.

`WriteError` does content negotiation to determine what to send back,
and for this reason, needs a value that can be marshaled as JSON. If
not, it will tend to write `"{}"` as the response body, which is
unhelpful to the client calling the API, and to the operator looking
in the logs.

To counter this, wrap any errors that aren't json.Marshaler in a
struct that is a json.Marshaler, so we can be sure the response body
will contain an error message.