Filter network stats to just 'eth0' interface, for consistency
This is done on the Monitor page already; the benefit is it avoids
showing whole-node stats (unless the user is on a really old Linux
where the main NIC is eth0).

The name 'eth0' and the fact there is only one interface is hard-coded
into Kubernetes; this approach will work for most people.

Ref: https://github.com/weaveworks/service-ui/issues/3801