Ensure top-level tags are applied on traces sampled via feature flag
Jaeger-client makes a sampling decision when the trace is created, and
any tags or child spans are ignored when not-sampled.

Fully resolving this is a chicken-and-egg problem in that we have to
look up user details before we see the `trace-debug-id` feature-flag,
but we can't do that until after the trace is created

As a simple improvement, move some `SetTag` calls later, and re-apply
the top-level method and url tags.