Use newer alpine instead of copying ca-certificates manually
The file we had checked in was so old that it didn't support let's
encrypt's root cert, instead relying on the cross-sign that expired
a few months ago. That's ridiculous. Just removing this step makes
that work.

They appear to have been checked in due to something-something, 2015
era connect to SNS - I don't believe that's still an issue.