users: Change test user behaviour to happen on login
Before, we could set `-direct-login` which replaced the email login
with a magic link we could click to log in. That's quite convenient.

We could also, at startup time, created a user `test@test.test` which
was an admin, and the owner of an org `Local Test Instance`.

This behaviour is all very neat, but what if instead we just let the
user log in however they want, and whatever they log in as/with, an
admin they become.