users: Encode the provider and the provider's login ID in sessions
If we're outsourcing our auth to a third party, we should keep track
of to whom we've outsourced it.

Note: this doesn't mean that the user is logged in to the provider in
question - our session time is 2 months and auto-renews forever, while
google provides sessions that last for maybe a day or two.