users: Introduce endpoint to verify session
Before, the frontend owned redirecting the user when the user got
logged out.

This introduces an endpoint for the frontend to redirect the user to
when their session expires, so they can be sent elsewhere. At the
moment this is a bit pointless, since we just send them back to the
/login path where they used to be sent to, but now that the endpoint
is there, I can move auth to auth0 with a backend-only change.