users: Stop auto-renewing the session cookie
I'd rather we didn't do this, as it disconnects us from the source of
authentication - e.g. we should check regularly with google if you're
still a user of that google account, or if you've e.g. left your role
and thus your google account behind.

At the moment, you've got a 2 month session time already, that's
enough.