users/db: Add method to retrieve a specific login
We list logins all the time, but we never actually just get
one. Concretely, we store the users's access key only in the login
entry in the database, so if I want to use the current access key for
the current login, I would have to list all logins and then filter for
a matching one - or by adding this method.