users: Forward query params from listLoginProviders
Before, listLoginProviders used to generate the complete link we
would use to login - including state, which contains CSRF tokens and
query parameters.

Service UI used this fact to forward query parameters through the
login process, in order to e.g. remember where a user was, or to make
GCP work. And while we still support this feature on the backend,
the link generation happens in verify, with listLoginProviders simply
returning a static link to it, so the forwarding chain broke.

Fix this by always including the incoming query parameters in the
response URLs in listLoginProviders.