users: Add backwards compatibility for github logins
Since I moved to auth0, all the old accounts needs to
re-associate, which they do automatically through the email address in
the case after this.  However, it seems a lot (a _lot_!) of github
accounts aren't returning emails, thus the next fallback fails, and
the user is denied access. In these cases, we could just, manually
extract the plain github user ID and look for a user that way - if we
find one, we can just let a new association be created, if not no harm
done.